Move common debug, memory & nvfile units to src/
