+++ src/tpm12/tpm_constants.h
@@ -143,12 +143,6 @@
   TPM buffer limits
 */
 
-/* This value is used to limit memory allocation to prevent resource overload. */
-
-#ifndef TPM_ALLOC_MAX
-#define TPM_ALLOC_MAX  0x20000  /* 128k bytes */
-#endif
-
 /* This is the increment by which the TPM_STORE_BUFFER grows.  A larger number saves realloc's.  A
    smaller number saves memory.
